Added fixes for packaging UI tests

+++ Tests/FeatherWidgets.TestUI.Arrangements/Packaging/StaticContent/Pages/ExportEditedPages.cs
@@ -45,7 +45,10 @@ namespace FeatherWidgets.TestUI.Arrangements
         public void ClearUp()
         {
             ServerOperations.Pages().DeleteAllPages();
-            ServerOperations.ModuleBuilder().DeleteDirectory(this.tempFolderPath);
+
+            if (System.IO.Directory.Exists(this.tempFolderPath))
+                ServerOperations.ModuleBuilder().DeleteDirectory(this.tempFolderPath);
+
             ServerOperations.Packaging().DeleteAllPackagesFromDB();
 
             for (int i = 0; i < ServerOperations.CustomFieldsNames().FieldNamesWithoutClassificationsEdited.Length; i++)
